View a list of all books
So here's our first, very simple feature! Listing books on a page.

I've written a feature spec which intentionally only tests the happy
path, i.e. when there are books to list. For the case of no books being
present, I've covered that in a view spec.

I don't normally write view specs, but I've used one here because I know
it's thoughtbot's approach, but also just an experiment for my own
learning. Would be interested to here other opinions on this approach.

I haven't styled anything yet, I thought it was better to get something
up so that the rest of the team can get started building.