Filter books by genre
https://trello.com/b/rxNcTk1k/ruby-bookshelf

I chose to use a Form Object so as to keep the filtering logic out of
the controller and the model.

I chose not to write a view spec for the `_filter` partial since it is
fairly simple without any loops or conditionals, and any problems should
be caught by the feature spec.

I also added a `seeds.db` so that developers can easily set up their
local instance with some sample data.