Adds user authentication
This introduces user authentication to the app, providing common
registration/login/logout endpoints, as well as a `current_user` method.

As discussed, authentication is implemented in terms of
`has_secure_password` in order to keep the system simple and reduce the
overhead of customizing an "out-of-the-box" solution, such as Clearance
or Devise.