Introduce Authentication module
This moves authentication methods from application controller into a
module. It also adjusts the `current_user` method not to return nil but
a new `Guest` instead.

In order to support the check for a user being
signed in, the `signed_in?` method was introduced, calling `present?` on
the return value of `current_user`. The `Guest` object will always
return false, by redefining `present?`.

While this approach is a quick and easy implementation,
the nature of reopening methods might lead to dangerous situations
further along. It should considered whether a different method should be
 added to both the `User` and `Guest` objects, under an API of our
control, that can better answer the question of a user being
`signed_in` or not.

This also includes slight changes to the sessions controller, namely
using `session.clear` instead of deleting only the `user_id` key on
logout. It also adjusts the class method signature style in the User
model.